chore: clean clang-tidy findings and retune config
Fix genuine issues (reserve before loops, explicit math parens, switch
default, narrowing cast). Retune .clang-tidy to bug-catching checks rather
than the full stylistic families (200+ no-value findings); keep
performance-enum-size disabled since those enums cross the C/Rust ABI as
int/i32. Add a bench-scoped .clang-tidy. clang-tidy now reports 0 across
the FFI, tests, and bench TUs.

+++ core/include/mape/compile_time.hpp
@@ -43,7 +43,7 @@ constexpr double bs_put(double spot, double strike, double rate, double vol,
 // ACT/365 year fraction between two integer day counts (never negative).
 constexpr double year_fraction(long start_day, long end_day) {
     const long days = end_day - start_day;
-    return (days > 0 ? days : 0) / 365.0;
+    return static_cast<double>(days > 0 ? days : 0) / 365.0;
 }
 
 // Continuously-compounded discount factor exp(-r * t), via the constexpr exp.
